split app into several files
